Modularize sessions and http
bacon: test
Resolves: OKTA-89655
