+++ lib/index.js
@@ -10,41 +10,24 @@
  * See the License for the specific language governing permissions and limitations under the License.
  */
 
-/* globals SDK_VERSION */
+/* globals STATE_TOKEN_COOKIE_NAME */
 /*jshint maxcomplexity:15 */
 
 require('./vendor/polyfills');
 
 var Q                 = require('q');
 var util              = require('./util');
+var http              = require('./http');
+var session           = require('./session');
 var cookies           = require('./cookies');
 var sdkCrypto         = require('./crypto');
 var ajaxRequest       = require('ajaxRequest'); // request type determined at build time
 var OAuthError        = require('./errors/OAuthError');
-var AuthApiError      = require('./errors/AuthApiError');
 var AuthSdkError      = require('./errors/AuthSdkError');
 var AuthPollStopError = require('./errors/AuthPollStopError');
 
-var STATE_TOKEN_COOKIE_NAME = 'oktaStateToken';
 var FRAME_ID = 'okta-oauth-helper-frame';
 var DEFAULT_POLLING_DELAY = 500;
-
-function getLink(obj, linkName, altName) {
-  if (!obj || !obj._links) {
-    return;
-  }
-  
-  var link = util.clone(obj._links[linkName]);
-
-  // If a link has a name and we have an altName, return if they match
-  if (link && link.name && altName) {
-    if (link.name === altName) {
-      return link;
-    }
-  } else {
-    return link;
-  }
-}
 
 function addStateToken(res, options) {
   var builtArgs = util.clone(options) || {};
@@ -80,7 +63,7 @@ function AuthTransaction(sdk, res) {
 }
 
 function postToTransaction(sdk, url, options) {
-  return post(sdk, url, options)
+  return http.post(sdk, url, options)
     .then(function(res) {
       return new AuthTransaction(sdk, res);
     });
@@ -93,9 +76,9 @@ function getPollFn(sdk, res, ref) {
     }
 
     // Get the poll function
-    var pollLink = getLink(res, 'next', 'poll');
+    var pollLink = util.getLink(res, 'next', 'poll');
     function pollFn() {
-      return post(sdk, pollLink.href, getStateToken(res), true, true);
+      return http.post(sdk, pollLink.href, getStateToken(res), true, true);
     }
 
     ref.isPolling = true;
@@ -176,7 +159,7 @@ function link2fn(sdk, res, obj, link, ref) {
 
       case 'GET':
         return function() {
-          return get(sdk, link.href);
+          return http.get(sdk, link.href);
         };
 
       case 'POST':
@@ -285,81 +268,6 @@ function flattenEmbedded(sdk, res, obj, ref) { /* jshint ignore:line */
   return obj;
 }
 
-// HTTP METHODS
-
-function httpRequest(sdk, url, method, args, dontSaveResponse) {
-  var headers = {
-    'Accept': 'application/json',
-    'Content-Type': 'application/json',
-    'X-Okta-SDK': 'okta-auth-js-' + SDK_VERSION
-  };
-  util.extend(headers, sdk.options.headers || {});
-
-  var options = {
-    headers: headers,
-    data: args || undefined
-  };
-
-  var err, res;
-  return new Q(sdk.options.ajaxRequest(method, url, options))
-    .then(function(resp) {
-      res = resp.responseText;
-      if (res && util.isString(res)) {
-        res = JSON.parse(res);
-      }
-
-      if (!dontSaveResponse) {
-        if (!res.stateToken) {
-          cookies.deleteCookie(STATE_TOKEN_COOKIE_NAME);
-        }
-      }
-
-      if (res && res.stateToken && res.expiresAt) {
-        cookies.setCookie(STATE_TOKEN_COOKIE_NAME, res.stateToken, res.expiresAt);
-      }
-
-      return res;
-    })
-    .fail(function(resp) { 
-      var serverErr = resp.responseText || {};
-      if (util.isString(serverErr)) {
-        try {
-          serverErr = JSON.parse(serverErr);
-        } catch (e) {
-          serverErr = {
-            errorSummary: 'Unknown error'
-          };
-        }
-      }
-
-      if (resp.status >= 500) {
-        serverErr.errorSummary = 'Unknown error';
-      }
-
-      if (sdk.options.transformErrorXHR) {
-        resp = sdk.options.transformErrorXHR(util.clone(resp));
-      }
-
-      err = new AuthApiError(serverErr, resp);
-
-      if (err.errorCode === 'E0000011') {
-        cookies.deleteCookie(STATE_TOKEN_COOKIE_NAME);
-      }
-
-      throw err;
-    });
-}
-
-function get(sdk, url, saveResponse) { /* jshint ignore:line */
-  url = util.isAbsoluteUrl(url) ? url : sdk.options.url + url;
-  return httpRequest(sdk, url, 'GET', undefined, !saveResponse);
-}
-
-function post(sdk, url, args, dontSaveResponse) { /* jshint ignore:line */
-  url = util.isAbsoluteUrl(url) ? url : sdk.options.url + url;
-  return httpRequest(sdk, url, 'POST', args, dontSaveResponse);
-}
-
 function OktaAuth(args) {
   var sdk = this;
 
@@ -392,11 +300,11 @@ function OktaAuth(args) {
   }
 
   sdk.session = {
-    close: util.bind(closeSession, sdk, sdk),
-    exists: util.bind(sessionExists, sdk, sdk),
-    get: util.bind(getSession, sdk, sdk),
-    refresh: util.bind(refreshSession, sdk, sdk),
-    setCookieAndRedirect: util.bind(setCookieAndRedirect, sdk, sdk)
+    close: util.bind(session.closeSession, sdk, sdk),
+    exists: util.bind(session.sessionExists, sdk, sdk),
+    get: util.bind(session.getSession, sdk, sdk),
+    refresh: util.bind(session.refreshSession, sdk, sdk),
+    setCookieAndRedirect: util.bind(session.setCookieAndRedirect, sdk, sdk)
   };
 
   sdk.tx = {
@@ -465,7 +373,7 @@ proto.verifyRecoveryToken = function (opts) {
 
 function transactionStatus(sdk, args) { /* jshint ignore:line */
   args = addStateToken(this, args);
-  return post(this, sdk.options.url + '/api/v1/authn', args);
+  return http.post(this, sdk.options.url + '/api/v1/authn', args);
 }
 
 function resumeTransaction(sdk, args) { /* jshint ignore:line */
@@ -491,7 +399,7 @@ function transactionExists(sdk) { /* jshint ignore:line */
 }
 
 function getWellKnown(sdk) {
-  return get(sdk, sdk.options.url + '/.well-known/openid-configuration');
+  return http.get(sdk, sdk.options.url + '/.well-known/openid-configuration');
 }
 
 function validateClaims(claims, iss, aud) {
@@ -578,7 +486,7 @@ function verifyIdToken(sdk, idToken, options) { /* jshint ignore:line */
 
   return getWellKnown(sdk)
     .then(function(res) {
-      return get(sdk, res['jwks_uri']);
+      return http.get(sdk, res['jwks_uri']);
     })
     .then(function(res) {
       var key = res.keys[0];
@@ -961,58 +869,6 @@ function getIdToken(sdk, oauthOptions, options) { /* jshint ignore:line */
     default:
       return Q.reject(new AuthSdkError('The full page redirect flow is not supported'));
   }
-}
-
-function sessionExists(sdk) { /* jshint ignore:line */
-  return sdk.session.get()
-    .then(function(res) {
-      if (res.status === 'ACTIVE') {
-        return true;
-      }
-      return false;
-    })
-    .fail(function() {
-      return false;
-    });
-}
-
-function getSession(sdk) { /* jshint ignore:line */
-  return get(sdk, '/api/v1/sessions/me')
-  .then(function(session) {
-    var res = util.omit(session, '_links');
-
-    res.refresh = function() {
-      return post(sdk, getLink(session, 'refresh').href);
-    };
-
-    res.user = function() {
-      return get(sdk, getLink(session, 'user').href);
-    };
-
-    return res;
-  })
-  .fail(function() {
-    // Return INACTIVE status on failure
-    return {status: 'INACTIVE'};
-  });
-}
-
-function closeSession(sdk) { /* jshint ignore:line */
-  return httpRequest(sdk, sdk.options.url + '/api/v1/sessions/me', 'DELETE', undefined, true);
-}
-
-function refreshSession(sdk) { /* jshint ignore:line */
-  return post(sdk, '/api/v1/sessions/me/lifecycle/refresh');
-}
-
-function setCookieAndRedirect(sdk, sessionToken, redirectUrl) { /* jshint ignore:line */
-  redirectUrl = redirectUrl || window.location.href;
-  window.location = sdk.options.url + '/login/sessionCookieRedirect' +
-    util.toQueryParams({
-      checkAccountSetupComplete: true,
-      token: sessionToken,
-      redirectUrl: redirectUrl
-    });
 }
 
 module.exports = OktaAuth;
