+++ lib/util.js
@@ -168,6 +168,23 @@ function find(collection, searchParams) {
   }
 }
 
+function getLink(obj, linkName, altName) {
+  if (!obj || !obj._links) {
+    return;
+  }
+  
+  var link = clone(obj._links[linkName]);
+
+  // If a link has a name and we have an altName, return if they match
+  if (link && link.name && altName) {
+    if (link.name === altName) {
+      return link;
+    }
+  } else {
+    return link;
+  }
+}
+
 module.exports = {
   base64UrlToBase64: base64UrlToBase64,
   base64UrlToString: base64UrlToString,
@@ -184,5 +201,6 @@ module.exports = {
   removeNils: removeNils,
   clone: clone,
   omit: omit,
-  find: find
+  find: find,
+  getLink: getLink
 };
